+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog][docs-changelog], and the version adher
 
 ## Unreleased
 ### Fixed
+- Rename InstanceStatus to InstanceState.
 - Crash when querying instance role and instance state.
 - Avoid parsing stderr within LocalExecutor class.